+++ app/src/main/java/kuis1/models/Cuboid.java
@@ -1,40 +1,40 @@
 package kuis1.models;
 
 public class Cuboid extends Object3D {
-  private int length;
-  private int width;
-  private int height;
+  private double length;
+  private double width;
+  private double height;
 
-  public Cuboid(int length, int width, int height) {
+  public Cuboid(double length, double width, double height) {
     this(ShapeName.CUBOID, length, width, height);
   }
 
-  public Cuboid(ShapeName shapeName, int length, int width, int height) {
+  public Cuboid(ShapeName shapeName, double length, double width, double height) {
     super(shapeName);
     this.length = length;
     this.width = width;
     this.height = height;
   }
 
-  public int getLength() {
+  public double getLength() {
     return length;
   }
 
-  public int getWidth() {
+  public double getWidth() {
     return width;
   }
 
-  public int getHeight() {
+  public double getHeight() {
     return height;
   }
 
   @Override
-  public int getVolume() {
+  public double getVolume() {
     return length * width * height;
   }
 
   @Override
-  public int getSurfaceArea() {
+  public double getSurfaceArea() {
     return 2 * (length * width + length * height + width * height);
   }
 }
