+++ app/src/main/java/kuis1/models/Cylinder.java
@@ -1,30 +1,30 @@
 package kuis1.models;
 
 public class Cylinder extends Object3D {
-  private int radius;
-  private int height;
+  private double radius;
+  private double height;
 
-  public Cylinder(int radius, int height) {
+  public Cylinder(double radius, double height) {
     super(ShapeName.CYLINDER);
     this.radius = radius;
     this.height = height;
   }
 
-  public int getRadius() {
+  public double getRadius() {
     return radius;
   }
 
-  public int getHeight() {
+  public double getHeight() {
     return height;
   }
 
   @Override
-  public int getVolume() {
-    return (int) (Math.PI * radius * radius * height);
+  public double getVolume() {
+    return Math.PI * radius * radius * height;
   }
 
   @Override
-  public int getSurfaceArea() {
-    return (int) (2.0 * Math.PI * radius * (radius + height));
+  public double getSurfaceArea() {
+    return 2.0 * Math.PI * radius * (radius + height);
   }
 }
