+++ app/src/main/java/kuis1/models/Object3D.java
@@ -34,7 +34,7 @@ public abstract class Object3D {
     }
   }
 
-  public abstract int getVolume();
+  public abstract double getVolume();
 
-  public abstract int getSurfaceArea();
+  public abstract double getSurfaceArea();
 }
